dedup x86 code
