+++ lib/Arch/Sleigh/Thumb2Arch.cpp
@@ -32,7 +32,8 @@ class SleighThumb2Arch final : public remill::sleigh::SleighArch {
  public:
   SleighThumb2Arch(llvm::LLVMContext *context_, OSName os_name_,
                    ArchName arch_name_)
-      : SleighArch(context_, os_name_, arch_name_, "ARM7_le.sla",
+      : ArchBase(context_, os_name_, arch_name_),
+        SleighArch(context_, os_name_, arch_name_, "ARM7_le.sla",
                    "ARMtTHUMB.pspec") {}
